feat(CLM-29402): Stop deleting thirdparty data for SBOM Manager licenses (#11024)

+++ insight-brain-service/src/main/java/com/sonatype/insight/brain/report/ReportService.java
@@ -167,7 +167,9 @@ public class ReportService
           .addAll(thirdPartyApplicationReportForInfrastructureAsCodeDTO.securityRows);
       includeThirdPartyData(tempFile, thirdPartyApplicationReportDTO);
       thirdPartyDataService.indexVulnerabilities(scanId);
-      thirdPartyDataService.deleteByScanId(scanId);
+      if (!productLicense.hasFeature(LicensedFeature.SBOM_MANAGER)) {
+        thirdPartyDataService.deleteByScanId(scanId);
+      }
     }
   }
